refactor(frontend): extract datastudio table meta editing
